requirements.txt
bottle
pylint
serpapi
python-dotenv
